requirements.txt
GitPython
GPUtil
PyYAML
setproctitle
tqdm
class-resolver
dgl-cu111
numpy
